[Style/#48]: 우물 상세 페이지 (타인)
