+++ src/app/(main)/well/[id]/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function WellPage() {
+  return <div>WellPage</div>;
+}
+
+export default WellPage;
