+++ src/app/(user)/[id]/well/[wellId]/page.tsx
@@ -1,7 +1,15 @@
+import { WellBookList, WellHeader, WellTitle } from '@/features/Well';
+import MainLayout from '@/layouts/MainLayout';
 import React from 'react';
 
 function UserWellDetailPage() {
-  return <div>UserWellDetailPage</div>;
+  return (
+    <MainLayout>
+      <WellHeader />
+      <WellTitle />
+      <WellBookList />
+    </MainLayout>
+  );
 }
 
 export default UserWellDetailPage;
