+++ src/features/Well/components/WellBook/WellBookList.tsx
@@ -17,6 +17,7 @@ function WellBookList() {
       animate='visible'
       variants={staggerContainerVariants}
     >
+      <div className='h-[12px] w-full shrink-0 rounded-t-[20px] bg-gray-900' />
       {hasNewReview && <WellBook />}
       <WellBook />
       <FrogOnBook />
