+++ src/features/Well/components/WellHeader.tsx
@@ -2,13 +2,15 @@
 
 import React from 'react';
 import { EditIcon } from 'public/icons';
-import { useSession } from 'next-auth/react';
 import BackButton from '@/components/Button/BackButton';
 import { useRouter } from 'next/navigation';
 
-function WellHeader() {
+interface Props {
+  hasEditButton?: boolean;
+}
+
+function WellHeader({ hasEditButton = false }: Props) {
   const router = useRouter();
-  const { data: session } = useSession();
 
   return (
     <div className='absolute left-[50%] top-0 z-10 flex w-[450px] translate-x-[-50%] gap-[20px] pt-[70px] mobile:left-0 mobile:w-full mobile:translate-x-0'>
@@ -17,7 +19,7 @@ function WellHeader() {
         fill='#B3B6C5'
         extraClass='absolute top-[28px] left-[28px] z-20'
       />
-      {session && (
+      {hasEditButton && (
         <button type='button' className='absolute right-[28px] top-[28px] z-20'>
           <EditIcon />
         </button>
