+++ src/features/Well/components/WellTitle.tsx
@@ -1,18 +1,15 @@
 'use client';
 
 import React from 'react';
-import { useSession } from 'next-auth/react';
-import Link from 'next/link';
 import { motion } from 'framer-motion';
-import { PlusIcon, ArrowIcon } from 'public/icons';
-import { PAGES } from '@/constants/page';
 import { IMAGES } from '@/constants/images';
 import Image from 'next/image';
-import Pointing from './Pointing';
 
-function WellTitle() {
-  const { data: session } = useSession();
+interface Props {
+  children?: React.ReactNode;
+}
 
+function WellTitle({ children }: Props) {
   return (
     <div className='flex-col-center relative flex w-full shrink-0'>
       <Image
@@ -29,26 +26,7 @@ function WellTitle() {
         className='flex flex-col gap-[20px] py-[50px]'
       >
         <h1 className='text-title-xl-bold'>소중한 나의 첫 우물</h1>
-        <div className='flex-col-center gap-[4px]'>
-          <div className='relative h-[28px] w-[28px]'>
-            <Link
-              href={session ? PAGES.SEARCH : PAGES.LANDING}
-              className='absolute inset-x-0 top-[50%] z-50 mx-auto -translate-y-1/2 cursor-pointer'
-            >
-              {session ? (
-                <PlusIcon />
-              ) : (
-                <ArrowIcon fill='#313239' width={28} height={28} />
-              )}
-            </Link>
-            <Pointing />
-          </div>
-
-          <h3 className='mt-[2px] text-body-xl-bold'>
-            {session ? '책 추가하기' : '로그인이 필요해요'}
-          </h3>
-          <span className='text-body-md text-gray-600'>현재 높이 0cm</span>
-        </div>
+        {children}
       </motion.div>
     </div>
   );
